SDK Overlays: Linux: disable explicit module builds
Explicit module builds crash the compiler when building the glibc
overlay library. The crash occurs in the Swift driver dependency
scanner. It appears to be related to vfs overlays, modulemaps,
injected header files, and explicit module builds.

This is odd though because the Android builds also have this
combination, but work.