[cxx-interop] Do not warn for implicit default constructors
The diagnostic that Swift was emitting (`This zero-initializes the backing memory of the struct ...`) was flaky, because it was only emitted when the default constructor of a C++ struct was instantiated by Swift. If it was instantiated by Clang, the diagnostic wasn't emitted. There is not a lot of value in it. Let's make the behavior consistent by removing the diagnostic.

rdar://161999293