Runtime: expose demangle() in Runtime module
We should finally expose the demangle functionality; It's been widely
used by calling into internal _swift_demangle and instead we should
offer a real API.

Previous discussions happened between 2019 and 2024, and just never
materialized in a complete implementation and proposal.

Right now, even more tools are in need of this API, as we are building
continious profiling solutions etc, so it is a good time to revisit
this topic.

This PR is roughly based off @Azoy's
https://github.com/swiftlang/swift/pull/25314/files#diff-fd379a721cc9a1c9ef6486eae713e945da842b42170d4d069029a57334371eba
from 2019, however it brings it over to the new Runtime module which is
a great place to put this functionality - even Backtrace had to recently
reinvent calling the demangling infra in this module.

Pending SE review.