PackSpecialization: Use indices from result map in replaceOriginalResults
We were previously using the indices of the results of the original function's
function convention, which is incorrect when the original function has a
non-pack result before its first pack result.

In this function, Pack{Int} is argument 0, but at index 1 of the result list:

sil ... () -> (Int, Pack{Int}) {
bb0(%0: $*Pack{Int}):
...
}

The previous code asserted our incorrect assumption that these two indices would
always be equal, so it worked in the cases where they were. We have removed the
assertion.