Add `Equatable` and `Hashable` conformance to `GUID`.
On Windows, `GUID` is a currency type and (along with its various typedefs) is
used pervasively Windows offers `IsEqualGUID()` and `UuidHash()` for comparing
and hashing them, respectively, but `IsEqualGUID()` is a macro in C mode and
`UuidHash()` only provides a 16-bit hash. We should provide conformance to these
protocols in the WinSDK overlay to provide equivalent functionality in Swift.