[AutoDiff][test] Enable some previously XFAIL'ed tests
In #84704, some tests were XFAIL'ed since they've become failing.
The root cause of failures is lack of ownership info on the 2nd run
of AutoDiff closure specialization pass. Ownership info is required for
the pass run after #84704, so at the moment only the 1st run of the pass
is effective.

Several test cases still remain passing because for some cases we are
lucky and all the inlining required for specialization is done before
the 1st pass run. This patch enables such test cases back so we have
at least some test coverage.