build.ps1: repair the Android SDK build
When building Android, we would fail to find the brotli build.
Explicitly pass the library paths and include paths to workaround the
issue. This is not ideal, but will unblock the build. Avoid the
installation of the library as it is not required, which avoids updating
the mtime on a null-build.