[SILGen]: ensure DI checks address-only let temporary allocations
Previously it was possible to reference uninitialized memory if a
closure that was part of a variable binding initializer referenced the
name of the uninitialized value. DI was ignoring these cases before
because no mark_uninitialized instruction was produced. Instead of
trying to skip emitting that instruction sometimes, just always emit it
if a temporary allocation is used so DI will catch liveness issues.