Fix warning -Wdangling-assignment-gsl in SIL/PrunedLiveness.h
Fix a potential use-after-free from a dangling pointer because of the
wrong usage of `llvm::function_ref` which does not own the function.
`std::function` is better suited in this case, or just use `auto` and
let compiler decide.

Fixes warning:
```
PrunedLiveness.h:994:10: warning: object backing the pointer 'op' will be destroyed at the end of the full-expression [-Wdangling-assignment-gsl]
  994 |     op = [](const SILInstruction *&ptr) -> SILBasicBlock *
```