utils: shuffle install order for toolchain dependencies
Install the testing dependencies first. This is primarily the python
dependencies. Doing this allows us to use the following arguments:
`-SkipBuild -SkipPackaging -Test lldb` to test LLDB without having to
rebuild the toolchain. This is helpful when trying to debug/diagnose
testing issues within LLDB without changing the implementation of LLDB.