[CS] Handle holes in `mergeEquivalenceClasses`
Make sure we preserve `TVO_CanBindToHole`. This unfortunately requires
introducing the concept of a "non-representative" hole for
representatives where the actual hole is given by another type variable
in the equivalence class. This avoids disturbing binding order and
regressing diagnostics. Once binding is less sensitive to ordering we
ought to be able to always choose the hole as the representative.