Use `__ehdr_start` on ELF instead of calling `dladdr()`.
This PR replaces our use of `__dso_handle` and (indirectly) `dladdr()` when
figuring out the base address of loaded Swift images on ELF-based platforms.
Instead, we use `__ehdr_start` which refers to the start of the image's ELF
header (i.e. the start of the file) and is provided by the compiler/linke for
this purpose.

This pointer is consumed by `swift_addNewDSOImage()` and, later, by Swift
Testing when running tests, but is otherwise unused, so the less work we do on
it at runtime, the better.