[Concurrency] Prevent task re-enqueues in continuation APIs
This explicitly adopts nonisolated nonsending funcs for the with
continuation APIs. This guarantees there are not spurious enqueues
between the caller and the closure executing, which is a requirement for
these methods.

This takes a risk on removing the bad #isolation based API from source,
however explicitly passing the isolation as something else than the
actual isolation would always have been incorrect, so we should try to
get this through.

There are further re-enqueue fixes to be done, but this specific API
should adopt the nonisolated nonsending approach.

Resolves rdar://162192512