[IRGen] Fix call generation to respect C calling conventions.
On 32-bit Windows, a number of different calling conventions are in use, but
Swift ignores them when generating call instructions in IRGen, and generates
calls that always use the default calling convention, `cdecl`.  This results in
stack corruption, because `stdcall` uses the `ret <nn>` instruction, which adds
its argument to the stack pointer on return.

rdar://163178024