[cxx-interop] Do not import typedef members of clang::RecordDecls
In C++, it is legal to have a typedef to an invalid type (e.g.,
a templated type that cannot be instantiated). We should not import
these eagerly in Swift, because doing so causes spurious type errors
that would not otherwise appear in C++---this is validated by the test
that is added in this patch, which runs clang.

Instead, we can rely on these types being lazily imported on-demand,
i.e., only when the type is referenced directly in Swift or indirectly
by an imported decl that is used in Swift.

rdar://145238539